refactor(admin): extract QR + force-reset modals from admin.js god file
Fixes #1589 (partial — highest-value safe slice). Continues the #1279
step-4b modularization by lifting two fully self-contained "modal wiring"
sections out of admin.js into www/js/admin/sections/:

- qr-modal.js: the tap-to-enlarge join-QR modal (openQRModal / closeQRModal
  / setupQRModal). Depends only on adminState + registerModalClose + the
  QRCode global — zero core-fn coupling, no DI.
- force-reset.js: the #777 emergency recovery modal (show/close/confirm +
  the _BEATIFY_LS_KEYS list). Depends only on registerModalClose + the
  BeatifyAuth global — no adminState, no core fns, no DI.

Strictly behavior-preserving: code moved verbatim. Init order, global
names and the WS-hub wiring are untouched. admin.js core imports the two
external entry points (setupQRModal/setupResetModal, plus openQRModal for
the home-view typeof-guarded handlers); closeQRModal and the reset
internals are now module-private. No window shims needed — every caller
stays inside the ESM bundle. admin.js: 3019 -> 2904 lines.

build:check passes (admin.min.js rebuilt, byte-for-byte), 373 vitest tests
green, eslint 0 errors.
